+++ changelog.md
@@ -1,5 +1,15 @@
 # UI Components Changelog
 
+## 1.1.0
+
+- Adding support for MapwizeForMapbox 2.0.0
+- Adding details in bottom view if the selected content has details to display
+- Adding subtitle in search result
+- Using navigation instead of direction if the user start from its current position
+- Fixing loading bar display
+- Displaying loading bar when searching for a direction
+- Improving design
+
 ## 1.0.4
 
 - Adding no result found in search result list